Static views clear contents on render
This is less performant than keeping around the old elements and reusing
them, but also way simpler.

Since we're following a design akin to React, meaning we render all of
our views based on the current model data, there should be no impact
from this removal.

If we have any dynamic views (those with control state) we will need to
do some more work to ensure that those state-holding elements are
reattached to the view on render (or never removed.)