Fixtures data is cleared between each test
This required us to no longer generate the Backbone.js model within the
fixtures files (as this was loaded only once).
Instead we do that within the fixture loader, which will re-run that
code if it detects that no fixtures have been loaded for that model yet,
for the given test.